pyproject.toml
[tool.poetry]
name = "ghostpost_backend"
version = "0.1.0"
description = "Backend portion using Django Rest Framework that interfaces with React FrontEnd running on the same machine for GhostPost."
authors = ["Reginald Tjitradi <rtjitradi@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.12.1"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
